[Fix & Feature] Login &  Forgot Password
Changed the layout of login.
Removed the "Check-In" option
Added a button for the link to "Forgot Password" feature.
Option to choose using email as recovery option or use phone number &
security
question and answer.
Security questions are currently  in the administration options panel so
that the admin can add/remove